+++ lib/ruuuby/class/str.rb
@@ -3,47 +3,55 @@
 # add various functions to existing class +String+ (and explicitly create aliases to play nice with IDEs)
 class ::String
 
-  # @param [*] str
+  # @param [String] str
+  #
+  # @raise [WrongParamType]
   #
   # @return [Boolean] true, if this string instance contains the provided str
   def ∋?(str)
-    raise ArgumentError.new("String#∌ included? operation only accepts a string argument, received a {#{str.class.to_s}} instead!") unless str.str?
+    raise Ruuuby::Err::param¬str(::String, 'str', str) unless str.str?
     self.include?(str)
   end
 
-  # @param [*] str
+  # @param [String] str
+  #
+  # @raise [WrongParamType]
   #
   # @return [Boolean] true, if this string instance contains the provided str
   def ∌?(str)
-    raise ArgumentError.new("String#∌ not-included? operation only accepts a string argument, received a {#{str.class.to_s}} instead!") unless str.str?
+    raise Ruuuby::Err::param¬str(::String, 'str', str) unless str.str?
     not self.include?(str)
   end
 
-  # @param [*] str
+  # @param [String] str
+  #
+  # @raise [WrongParamType]
   #
   # @return [Boolean] true, if this string instance is not contained in the provided str
   def ∉?(str)
-    raise ArgumentError.new("String#∉ not-contained? operation only accepts a string argument, received a {#{str.class.to_s}} instead!") unless str.str?
+    raise Ruuuby::Err::param¬str(::String, 'str', str) unless str.str?
     str.∌?(self)
   end
 
-  # @param [*] str
+  # @param [String] str
+  #
+  # @raise [WrongParamType]
   #
   # @return [Boolean] true, if this string instance is contained in the provided str
   def ∈?(str)
-    raise ArgumentError.new("String#∈ contained? operation only accepts a string argument, received a {#{str.class.to_s}} instead!") unless str.str?
+    raise Ruuuby::Err::param¬str(::String, 'str', str) unless str.str?
     str.∋?(self)
   end
 
-  # @param [String] start the text that this string start with
+  # @param [String]  start the text that this string start with
   # @param [Boolean] use_partial_fill_in ex: 'baa'.ensure_start!('bba', true) -> 'bbaa', 'baa'.ensure_start!('bba', false) -> 'bbabaa'
   #
-  # @raise [ArgumentError] thrown if the parameter provided is not a +String+ instance
+  # @raise [WrongParamType]
   #
   # @return [String] the original string with the starting text added if at least partially missing prior
   def ensure_start!(start, use_partial_fill_in)
-    raise ArgumentError.new("Provided argument is a {#{start.class.to_s}}, not a String!") unless start.str?
-    raise ArgumentError.new("Provided argument is a {#{use_partial_fill_in.class.to_s}}, not a Boolean!") unless use_partial_fill_in.bool?
+    raise Ruuuby::Err::param¬str(::String, 'start', start) unless start.str?
+    raise Ruuuby::Err::param¬bool(::String, 'use_partial_fill_in', use_partial_fill_in) unless use_partial_fill_in.bool?
     return self if start.length == 0 || self.start_with?(start)
     return self >> start if (self.length == 0 || !use_partial_fill_in)
     last_matched = ''
@@ -59,12 +67,12 @@ class ::String
   # @param [String] ending the text that this string should end with
   # @param [Boolean] use_partial_fill_in ex: 'baa'.ensure_ending!('aab', true) -> 'baab', 'baa'.ensure_ending!('aab', false) -> 'baaaab'
   #
-  # @raise [ArgumentError] thrown if the parameter provided is not a String instance
+  # @raise [WrongParamType]
   #
   # @return [String] the original string with ending-string added if missing prior
   def ensure_ending!(ending, use_partial_fill_in)
-    raise ArgumentError.new("Provided argument is a{#{ending.class.to_s}}, not a String!") unless ending.str?
-    raise ArgumentError.new("Provided argument is a{#{use_partial_fill_in.class.to_s}}, not a Boolean!") unless use_partial_fill_in.bool?
+    raise Ruuuby::Err::param¬str(::String, 'ending', ending) unless ending.str?
+    raise Ruuuby::Err::param¬bool(::String, 'use_partial_fill_in', use_partial_fill_in) unless use_partial_fill_in.bool?
     return self if ending.length == 0 || self.end_with?(ending)
     return self << ending if self.length == 0 || !use_partial_fill_in
     last_matched = ''
@@ -76,4 +84,5 @@ class ::String
     end
     self << (last_matched == '' ? ending : ending[last_matched.length..ending.length-1])
   end
+
 end