+++ spec/class/hsh_spec.rb
@@ -4,8 +4,10 @@ RSpec.describe 'hsh' do
 
   context 'extends class[Hash]' do
     it 'by creating needed aliases' do
-      expect_func_alias(::Hash, :key?, :🔑?)
-      expect_func_alias(::Hash, :key?, :∃🔑?)
+      expect(::Hash.∃func_alias?(:key?, :🔑?)).to eq(true)
+      expect(::Hash.∃func_alias?(:key?, :∃🔑?)).to eq(true)
+      expect({}.class.∃func_alias?(:key?, :🔑?)).to eq(true)
+      expect({}.class.∃func_alias?(:key?, :∃🔑?)).to eq(true)
     end
   end